requirements.txt
python-chess
matplotlib
seaborn
stockfish
cairosvg
numpy
imageio
imageio-ffmpeg
Pillow
polyglot
